refactor: migrate to module-builder
