+++ package.json
@@ -5,25 +5,27 @@
   "repository": "nuxt-community/tailwindcss-module",
   "license": "MIT",
   "exports": {
-    ".": "./dist/index.js",
-    "./src/runtime/*": "./dist/runtime/*",
-    "./package.json": "./package.json"
+    ".": {
+      "require": "./dist/module.cjs",
+      "import": "./dist/module.mjs"
+    }
   },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./dist/module.cjs",
+  "types": "./dist/types.d.ts",
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "siroc build",
-    "dev": "nuxt example",
+    "build": "nuxt-module-build",
+    "dev": "nuxt dev playground",
     "docs": "nuxt docs",
     "lint": "eslint --ext .js,.ts,.vue",
-    "test": "yarn lint && jest",
     "prepack": "yarn build",
-    "release": "yarn test && standard-version && git push --follow-tags && npm publish"
+    "release": "yarn test && standard-version && git push --follow-tags && npm publish",
+    "test": "yarn lint && jest"
   },
   "dependencies": {
+    "@nuxt/kit": "npm:@nuxt/kit-edge@latest",
     "@nuxt/postcss8": "^1.1.3",
     "autoprefixer": "^10.4.2",
     "chalk": "^4.1.2",
@@ -38,6 +40,7 @@
     "ufo": "^0.7.10"
   },
   "devDependencies": {
+    "@nuxt/module-builder": "latest",
     "@nuxt/test-utils": "latest",
     "@nuxtjs/eslint-config-typescript": "latest",
     "@types/jest": "latest",
@@ -45,7 +48,6 @@
     "eslint": "latest",
     "jest": "latest",
     "nuxt-edge": "latest",
-    "siroc": "latest",
     "standard-version": "latest"
   }
 }
